ovis1.6-llama3.2-3b/requirements.txt
torch #==2.2.0
transformers #==4.44.2
numpy #==1.24.3
pillow #==10.3.0
